Apply https://github.com/evanmiller/libxls/commit/c7e5e49518027c2a8bf65c933b1a7f8804a8f7fb

+++ src/libxls/xlstool.h
@@ -49,6 +49,6 @@ extern void xls_showCell(struct st_cell_data* cell);
 extern void xls_showFont(struct st_font_data* font);
 extern void xls_showXF(XF8* xf);
 extern void xls_showFormat(struct st_format_data* format);
-extern char* xls_getfcell(xlsWorkBook* pWB, struct st_cell_data* cell, WORD *label);
+extern char* xls_getfcell(xlsWorkBook* pWB, struct st_cell_data* cell, BYTE *label);
 extern char* xls_getCSS(xlsWorkBook* pWB);
 extern void xls_showBOF(BOF* bof);
